refactor: rename plugin and namespace `rich` -> `richgeo`
