+++ src/services/CMakeLists.txt
@@ -1,6 +1,6 @@
 add_subdirectory(geometry/dd4hep)
 add_subdirectory(geometry/acts)
-add_subdirectory(geometry/rich)
+add_subdirectory(geometry/richgeo)
 add_subdirectory(io/podio)
 add_subdirectory(randomgenerator)
 add_subdirectory(log)
