+++ src/services/geometry/richgeo/ActsGeo.h
@@ -21,7 +21,7 @@
 // local
 #include "RichGeo.h"
 
-namespace rich {
+namespace richgeo {
   class ActsGeo {
     public:
