+++ src/services/geometry/richgeo/RichGeo_service.h
@@ -31,8 +31,8 @@ class RichGeo_service : public JService {
     ~RichGeo_service();
 
     // return pointers to geometry bindings; initializes the bindings upon the first time called
-    rich::IrtGeo *GetIrtGeo(std::string detector_name);
-    rich::ActsGeo *GetActsGeo(std::string detector_name);
+    richgeo::IrtGeo *GetIrtGeo(std::string detector_name);
+    richgeo::ActsGeo *GetActsGeo(std::string detector_name);
 
   private:
     RichGeo_service() = default;
@@ -42,8 +42,8 @@ class RichGeo_service : public JService {
     std::once_flag   m_init_acts;
     JApplication     *m_app       = nullptr;
     dd4hep::Detector *m_dd4hepGeo = nullptr;
-    rich::IrtGeo     *m_irtGeo    = nullptr;
-    rich::ActsGeo    *m_actsGeo   = nullptr;
+    richgeo::IrtGeo  *m_irtGeo    = nullptr;
+    richgeo::ActsGeo *m_actsGeo   = nullptr;
 
     std::shared_ptr<spdlog::logger> m_log;
     bool m_verbose;
